Rename InputFieldValidator move to TestDoubles namespace
